pyproject.toml
[project]
name = "travel-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.1.8",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "pydantic>=2.0.0",
    "langchain-deepseek>=1.0.1",
    "python-dotenv>=1.2.2",
    "cryptography>=42.0.0",
    "pyjwt>=2.8.0",
]
